chore: Cleanup refactor
