+++ packages/ref-metrics-pmp/src/cmip_ref_metrics_pmp/pmp_driver.py
@@ -1,9 +1,102 @@
 import importlib.metadata
 import importlib.resources
+import json
 import pathlib
 import subprocess
+from typing import Any
+
+from cmip_ref_core.pycmec.metric import CMECMetric
+from cmip_ref_core.pycmec.output import CMECOutput
 
 DEFAULT_CONDA_ENV = "ref-metrics-pmp"
+
+
+def _remove_nested_key(data: dict[str, Any], key: str) -> dict[str, Any]:
+    """
+    Remove a nested key from a dictionary
+
+    Parameters
+    ----------
+    data
+        Dictionary to remove the key from
+    key
+        Key to remove
+
+    Returns
+    -------
+        The dictionary with the key removed
+    """
+    if key in data:
+        data.pop(key)
+    for k, v in data.items():
+        if isinstance(v, dict):
+            data[k] = _remove_nested_key(v, key)
+    return data
+
+
+def process_json_result(
+    json_filename: pathlib.Path, png_files: list[pathlib.Path], data_files: list[pathlib.Path]
+) -> tuple[CMECOutput, CMECMetric]:
+    """
+    Process a PMP JSON result into the appropriate CMEC bundles
+
+    Parameters
+    ----------
+    json_filename
+        Filename of the JSON file that is written out by PMP
+    png_files
+        List of PNG files to be included in the output
+    data_files
+        List of data files to be included in the output
+
+    Returns
+    -------
+        tuple of CMEC output and metric bundles
+    """
+    with open(json_filename) as fh:
+        json_result = json.load(fh)
+
+    cmec_output = CMECOutput.create_template()
+    cmec_output["provenance"] = {**cmec_output["provenance"], **json_result["provenance"]}
+
+    # Add the plots and data files
+    for fname in png_files:
+        cmec_output["plots"][fname.name] = {
+            "filename": str(fname),
+            "long_name": "Plot",
+            "description": "Plot produced by the metric",
+        }
+    for fname in data_files:
+        cmec_output["data"][fname.name] = {
+            "filename": str(fname),
+            "long_name": "Output data",
+            "description": "Data produced by the metric",
+        }
+
+    cmec_metric = CMECMetric.create_template()
+    cmec_metric["DIMENSIONS"] = {}
+    dimensions = json_result["DIMENSIONS"]
+
+    if "dimensions" in dimensions:
+        # Merge the contents of inner "dimensions" into the parent "DIMENSIONS"
+        dimensions.update(dimensions["dimensions"])
+        del dimensions["dimensions"]
+
+    if "statistic" in dimensions["json_structure"]:
+        dimensions["json_structure"].remove("statistic")
+        dimensions.pop("statistic")
+
+    # Remove the "attributes" key from the RESULTS
+    # This isn't standard CMEC output, but it is what PMP produces
+    results = _remove_nested_key(json_result["RESULTS"], "attributes")
+
+    cmec_metric["RESULTS"] = results
+    cmec_metric["DIMENSIONS"] = dimensions
+
+    if "provenance" in json_result:
+        cmec_metric["provenance"] = json_result["provenance"]
+
+    return CMECOutput(**cmec_output), CMECMetric(**cmec_metric)
 
 
 def get_resource_filename(package: str, resource_name: str | pathlib.Path, use_resources: bool) -> str:
